fix load data in all device
